requirements.txt
django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
drf-spectacular==0.27.2
python-dotenv==1.0.1
psycopg[binary]>=3.1
gunicorn==22.0.0
django-cors-headers==4.4.0
djangorestframework_simplejwt==5.5.1
django-cors-headers
django-extensions==4.1
openpyxl==3.1.5
requests==2.32.3
